+++ frontend/src/test/java/org/abs_models/backend/common/StmtTests.java
@@ -47,13 +47,13 @@ public class StmtTests extends SemanticTests {
     }
     
     @Test
-    public void declIfExpStmt() {
+    public void declWhenExpStmt() {
         // dynamic Java backend generated invalid code for this case
         assertEvalTrue("{Bool x = True; Int i = when x then 1 else 2; Bool testresult = (i == 1); }");
     }
 
     @Test
-    public void assignIfExpStmt() {
+    public void assignWhenExpStmt() {
         // dynamic Java backend generated invalid code for this case
         assertEvalTrue("{Bool x = True; Int i = 0; i = when x then 1 else 2; Bool testresult = (i == 1); }");
     }