+++ CHANGELOG.md
@@ -10,6 +10,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- The toolchain now uses Xtext (https://eclipse.org/Xtext) as parsing framework.  This switch caused some minor syntax changes, listed below.
+
+  - Incompatible change: the conditional expression now uses `when` as keyword instead of `if`.  (The conditional statement uses `if` as before.)
+
+  - Incompatible change: the pattern matching statement now uses `switch` as keyword instead of `case`.  (The pattern-matching expression uses `case` as before.)
+
 ### Removed
 
 ### Fixed